Rename Misc.ts to UrlUtils.ts
